renamed layouts folder to components
