+++ src/App.js
@@ -2,12 +2,11 @@ import React from 'react';
 import logo from './logo.svg';
 import { Counter } from './features/counter/Counter';
 import './App.css';
-import Sidebar from './Layouts/Sidebar';
+import Sidebar from './components/Sidebar';
 
 function App() {
   return (
     <div className="App">
-      {/* <MiniDrawer /> */}
       <Sidebar />
     </div>
   );
