+++ src/components/MiniDrawer.js
@@ -1,30 +1,29 @@
-import React from 'react';
-import clsx from 'clsx';
-import { makeStyles, useTheme } from '@material-ui/core/styles';
-import Drawer from '@material-ui/core/Drawer';
-import AppBar from '@material-ui/core/AppBar';
-import Toolbar from '@material-ui/core/Toolbar';
-import List from '@material-ui/core/List';
-import CssBaseline from '@material-ui/core/CssBaseline';
-import Typography from '@material-ui/core/Typography';
-import Divider from '@material-ui/core/Divider';
-import IconButton from '@material-ui/core/IconButton';
-import MenuIcon from '@material-ui/icons/Menu';
-import ChevronLeftIcon from '@material-ui/icons/ChevronLeft';
-import ChevronRightIcon from '@material-ui/icons/ChevronRight';
-import ListItem from '@material-ui/core/ListItem';
-import ListItemIcon from '@material-ui/core/ListItemIcon';
-import ListItemText from '@material-ui/core/ListItemText';
-import Button from '@material-ui/core/Button';
+import React from './node_modules/react';
+import clsx from './node_modules/clsx';
+import { makeStyles, useTheme } from './node_modules/@material-ui/core/styles';
+import Drawer from './node_modules/@material-ui/core/Drawer';
+import AppBar from './node_modules/@material-ui/core/AppBar';
+import Toolbar from './node_modules/@material-ui/core/Toolbar';
+import List from './node_modules/@material-ui/core/List';
+import CssBaseline from './node_modules/@material-ui/core/CssBaseline';
+import Typography from './node_modules/@material-ui/core/Typography';
+import Divider from './node_modules/@material-ui/core/Divider';
+import IconButton from './node_modules/@material-ui/core/IconButton';
+import MenuIcon from './node_modules/@material-ui/icons/Menu';
+import ChevronLeftIcon from './node_modules/@material-ui/icons/ChevronLeft';
+import ChevronRightIcon from './node_modules/@material-ui/icons/ChevronRight';
+import ListItem from './node_modules/@material-ui/core/ListItem';
+import ListItemIcon from './node_modules/@material-ui/core/ListItemIcon';
+import ListItemText from './node_modules/@material-ui/core/ListItemText';
+import Button from './node_modules/@material-ui/core/Button';
 
-import AccountCircleIcon from '@material-ui/icons/AccountCircle';
-import AudiotrackIcon from '@material-ui/icons/Audiotrack';
-import HeadsetIcon from '@material-ui/icons/Headset';
-import MicIcon from '@material-ui/icons/Mic';
-import FavoriteIcon from '@material-ui/icons/Favorite';
-import PublicIcon from '@material-ui/icons/Public';
-import QueueMusicIcon from '@material-ui/icons/QueueMusic';
-import PlayCircleFilledIcon from '@material-ui/icons/PlayCircleFilled';
+import AccountCircleIcon from './node_modules/@material-ui/icons/AccountCircle';
+import AudiotrackIcon from './node_modules/@material-ui/icons/Audiotrack';
+import HeadsetIcon from './node_modules/@material-ui/icons/Headset';
+import MicIcon from './node_modules/@material-ui/icons/Mic';
+import FavoriteIcon from './node_modules/@material-ui/icons/Favorite';
+import QueueMusicIcon from './node_modules/@material-ui/icons/QueueMusic';
+import PlayCircleFilledIcon from './node_modules/@material-ui/icons/PlayCircleFilled';
 
 
 const drawerWidth = 240;
